backend/requirements.txt
# Core dependencies - install these first
numpy==1.26.4
cython==3.0.10

# Then install the rest
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.5.2
sentence-transformers==2.7.0
# faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.8.1
beautifulsoup4==4.12.3
# lxml==5.2.1
markdown==3.6
chardet==5.2.0
Pillow
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
flask-socketio==5.3.6
python-socketio==5.11.2
eventlet==0.35.2
gunicorn==22.0.0
psycopg==3.1.18
